Corrected ImmutablePlatformModule according to review.
